feat(emprinten): Flip sandbox function call check logic around.
Jinja2 allows function calls unless they are marked with unsafe_callable
or alters_data attribute. The latter is set by Django on default db
altering functions, but it is not automatically used on other functions
that call those.

For now, allow only functions marked with is_safe_to_call to be called.