feat(emprinten): Add view for rendering pdf from csv data
PDF Projects can now be added to an Event so that event personnel
(defined with CBAC rules) may access the feature. This can be used to
create either multiple PDF's (in a zip file) or a repeating PDF from a
csv dataset.

- project_index shows just a form with file field and displays the list
of variables referred in the used template.
- handle_upload is the form action that actually does the pdf rendering.
It doesn't save the form file anywhere and relies that django will clean
up its own temp files (if any, in case in-memory file was not enough).